feat: migrate app bootstrapping to multitargeted project
