+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.Base/base.props
@@ -10,7 +10,5 @@
 				XamlRuntime="WinUI"
 				DependentUpon="App.xaml"
 				Link="App.xaml.cs" />
-		<Compile Include="$(MSBuildThisFileDirectory)GlobalUsings.cs"
-				Link="GlobalUsings.cs" />
 	</ItemGroup>
 </Project>
